+++ artemis/artemis-core/src/main/java/org/eclipse/jnosql/artemis/repository/DynamicReturn.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection.repository;
+package org.eclipse.jnosql.artemis.repository;
 
 import jakarta.nosql.NonUniqueResultException;
 import jakarta.nosql.Sort;
