+++ artemis/artemis-core/src/main/java/org/eclipse/jnosql/artemis/repository/PaginationDynamicExecutorQueryConverter.java
@@ -12,11 +12,11 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection.repository;
+package org.eclipse.jnosql.artemis.repository;
 
-import jakarta.nosql.mapping.DynamicQueryException;
 import jakarta.nosql.mapping.Page;
 
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
@@ -25,56 +25,55 @@ import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-
 /**
- * The default implementation to {@link DynamicExecutorQueryConverter}.
- * This implementation won't work with pagination
+ * An implementation of {@link DynamicExecutorQueryConverter} that uses pagination at it.
  */
-final class DefaultDynamicExecutorQueryConverter implements DynamicExecutorQueryConverter {
+final class PaginationDynamicExecutorQueryConverter implements DynamicExecutorQueryConverter {
+
 
     @Override
     public <T> T toInstance(DynamicReturn<T> dynamic) {
-        Optional<T> optional = dynamic.singleResult();
+        Optional<T> optional = dynamic.singleResultPagination();
         return optional.orElse(null);
     }
 
     @Override
     public <T> Optional<T> toOptional(DynamicReturn<T> dynamic) {
-        return dynamic.singleResult();
+        return dynamic.singleResultPagination();
     }
 
     @Override
     public <T> List<T> toList(DynamicReturn<T> dynamic) {
-        return dynamic.result().collect(Collectors.toList());
+        return dynamic.streamPagination().collect(Collectors.toList());
     }
 
     @Override
     public <T> Set<T> toSet(DynamicReturn<T> dynamic) {
-        return dynamic.result().collect(Collectors.toSet());
+        return dynamic.streamPagination().collect(Collectors.toCollection(HashSet::new));
     }
 
     @Override
     public <T> LinkedList<T> toLinkedList(DynamicReturn<T> dynamic) {
-        return dynamic.result().collect(Collectors.toCollection(LinkedList::new));
+        return dynamic.streamPagination().collect(Collectors.toCollection(LinkedList::new));
     }
 
     @Override
     public <T> Stream<T> toStream(DynamicReturn<T> dynamic) {
-        return dynamic.result();
+        return dynamic.streamPagination();
     }
 
     @Override
     public <T> TreeSet<T> toTreeSet(DynamicReturn<T> dynamic) {
-        return dynamic.result().collect(Collectors.toCollection(TreeSet::new));
+        return dynamic.streamPagination().collect(Collectors.toCollection(TreeSet::new));
     }
 
     @Override
     public <T> Object toDefault(DynamicReturn<T> dynamic) {
-        return dynamic.result();
+        return dynamic.streamPagination();
     }
 
     @Override
     public <T> Page<T> toPage(DynamicReturn<T> dynamic) {
-        throw new DynamicQueryException("There is not pagination at the method: " + dynamic.getMethod());
+        return dynamic.getPage();
     }
 }
