+++ artemis/artemis-core/src/test/java/org/eclipse/jnosql/artemis/repository/PaginationDynamicExecutorQueryConverterTest.java
@@ -12,13 +12,13 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection.repository;
+package org.eclipse.jnosql.artemis.repository;
 
 import jakarta.nosql.mapping.Page;
 import jakarta.nosql.mapping.Pagination;
-import org.eclipse.jnosql.artemis.reflection.repository.DynamicExecutorQueryConverter;
-import org.eclipse.jnosql.artemis.reflection.repository.DynamicReturn;
-import org.eclipse.jnosql.artemis.reflection.repository.PaginationDynamicExecutorQueryConverter;
+import org.eclipse.jnosql.artemis.repository.DynamicExecutorQueryConverter;
+import org.eclipse.jnosql.artemis.repository.DynamicReturn;
+import org.eclipse.jnosql.artemis.repository.PaginationDynamicExecutorQueryConverter;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
